Increase umac block count to 64 bits. Some reordering of umac contexts.

+++ umac-l2.c
@@ -57,7 +57,7 @@ _umac_l2_init (unsigned size, uint32_t *k)
 
 void
 _umac_l2(const uint32_t *key, uint64_t *state, unsigned n,
-	 unsigned count, uint64_t *prev, const uint64_t *m)
+	 uint64_t count, uint64_t *prev, const uint64_t *m)
 {
   unsigned i;
 
@@ -95,7 +95,7 @@ _umac_l2(const uint32_t *key, uint64_t *state, unsigned n,
 
 void
 _umac_l2_final(const uint32_t *key, uint64_t *state, unsigned n,
-	       unsigned count, uint64_t *prev)
+	       uint64_t count, uint64_t *prev)
 {
   unsigned i;
